MISC address @mbondels comments
